:heavy_plus_sign: Add Navigation and CodeFix

+++ app/src/main/java/com/ashish/nowplayingapp/model/Movie.kt
@@ -5,12 +5,12 @@ data class Movie(
     val id : Long,
     val backdrop_path: String,
     val original_title: String,
-    val popularity: Long,
+    val popularity: Float,
     val vote_average : Float
 ){
     companion object{
         fun sampleMovie() = Movie(
-            34435,"h31SOVlekuHXsMWVGxI8nPPfY82.jpg","The Inception " , 54957 , 3.7f
+            34435,"h31SOVlekuHXsMWVGxI8nPPfY82.jpg","The Inception " , 54957F , 3.7f
         )
     }
 }
